move cw blocks to cw example
